chargement données du panier OK

+++ front/js/cart.js
@@ -1,26 +1,31 @@
 let panier = []
 if (localStorage.length > 0) {
-    let record = {}
-    for (let cpt=0; cpt < localStorage.length; cpt++) {
-        const items = JSON.parse(localStorage.getItem(localStorage.key(cpt)))
-        const itemId        = items.id
-        const itemColor     = items.color
-        const itemQuantity  = items.quantity
-        c(itemId + " - " + itemColor + " - " + itemQuantity)
-        getProduct(localStorage.key(cpt)).then((data) => {
+    let record      = {}
+    const numberOfRecord = localStorage.length 
+    for (let cpt=0; cpt < numberOfRecord; cpt++) {
+        const items         = JSON.parse(localStorage.getItem(localStorage.key(cpt)))
+        // const itemId        = items.id
+        // const itemColor     = items.color
+        // const itemQuantity  = items.quantity
+        // c(itemId + " / " + itemColor + " / " + itemQuantity)
+        fetch("http://localhost:3000/api/products/" + items.id)  
+        .then((res) => res.json())
+        .then((data) => { 
             record = {
-                id          : itemId,
-                color       : itemColor,
-                quantity    : itemQuantity,
+                id          : items.id,
+                color       : items.color,
+                quantity    : items.quantity,
                 price       : data.price,
                 name        : data.name,
                 description : data.description,
                 imageUrl    : data.imageUrl,
                 altTxt      : data.altTxt,
             }
+            c(record)
             panier.push(record)
             createStructure(record)
-        })
+         })
+        .catch((error) => { window.alert("Connexion au serveur impossible !") })
     }
     displayQuantityTotal()
     displayTotalPrice()
@@ -32,43 +37,50 @@ if (localStorage.length > 0) {
 function getProduct(varId) {
     return fetch("http://localhost:3000/api/products/" + varId)  
         .then((res) => res.json())
-        .then((data) => { return data })
-        .catch((error) => {
-            window.alert("Connexion au serveur impossible !")
-          })
+        .then((data) => { return JSON.parse(data) })
+        .catch((error) => { window.alert("Connexion au serveur impossible !") })
 }
 
 function createStructure(item) {
-    c(item)
     const parent = document.querySelector("#cart__items")
-    parent.innerHTML += `
-    <article class="cart__item" data-id="${item.id}" data-color="${item.color}">
-        <div class="cart__item__img">
-            <img src="${item.imageUrl}" alt="${item.altTxt}">
-        </div>
-        <div class="cart__item__content">
-            <div class="cart__item__content__description">
-                <h2>${item.name}</h2>
-                <p>${item.color}</p>
-                <p>${item.price} €</p>
+    if (testParent(parent)) {
+        parent.innerHTML += `
+        <article class="cart__item" data-id="${item.id}" data-color="${item.color}">
+            <div class="cart__item__img">
+                <img src="${item.imageUrl}" alt="${item.altTxt}">
             </div>
-            <div class="cart__item__content__settings">
-                <div class="cart__item__content__settings__quantity">
-                    <p>Qté : </p>
-                    <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${item.quantity}">
+            <div class="cart__item__content">
+                <div class="cart__item__content__description">
+                    <h2>${item.name}</h2>
+                    <p>${item.color}</p>
+                    <p>${item.price} €</p>
                 </div>
-                <div class="cart__item__content__settings__delete">
-                    <p class="deleteItem">Supprimer</p>
+                <div class="cart__item__content__settings">
+                    <div class="cart__item__content__settings__quantity">
+                        <p>Qté : </p>
+                        <input type="number" class="itemQuantity" name="itemQuantity" min="1" max="100" value="${item.quantity}">
+                    </div>
+                    <div class="cart__item__content__settings__delete">
+                        <p class="deleteItem">Supprimer</p>
+                    </div>
                 </div>
             </div>
-        </div>
-    </article>`
-    const itemSearch = `article[data-id="${item.id}"][data-color="${item.color}"]`
-    const varInput = document.querySelector(itemSearch)
-    varInput.addEventListener("input", () => updateQuantity(item.id, input.value, item))
+        </article>`
+        const itemSearch = `article[data-id="${item.id}"][data-color="${item.color}"]`
+        const varInput = document.querySelector(itemSearch)
+        varInput.addEventListener("input", () => updateQuantity(item.id, input.value, item))
+    }
 }
 
 function c(texte) { console.log(texte) }
+
+
+// fonction qui test le retour d'un querySelector
+function testParent(parent) {
+    if (parent != null) return true
+    else return false
+}
+
 
 function updateQuantity(id, newValue, item) {
     const itemToUpdate = panier.find((item) => item.id === id)
